Make package index template more minimal
The original template was specifically configured for the Arduino SAMD Boards platform. This will need to be customized
for each individual platform the "Platform Release" workflow is added to. It seems best to leave a portion of the SAMD
configuration in place as an example, but there's no need for the full thing.