Remove dependencies caching from workflows
Doubts were raised about whether the caching I had previously added to the workflows had any benefits. I don't know
enough about this subject to make a call on that and I have been unable to get any more information on the subject.

Since the caching significantly increases the complexity of the workflows, which may make them more difficult to maintain
and contribute to, I think it's best to just remove all the caching for now. I hope to eventually be able to revisit this topic and restore caching wherever it is definitely beneficial.

Background:
GitHub Provides the `actions/cache` action for caching workflow dependencies and build output:
https://github.com/actions/cache

There is a tutorial on it here:
https://docs.github.com/en/actions/guides/caching-dependencies-to-speed-up-workflows
Which says:

> Jobs on GitHub-hosted runners start in a clean virtual environment and must download dependencies each time, causing
increased network utilization, longer runtime, and increased cost. To help speed up the time it takes to recreate these
files, GitHub can cache dependencies you frequently use in workflows.

So my thought was that it would be a good idea to set up dependencies caching in all the workflow templates. I have long
thought about this subject as it relates to the sketch compilation workflows. On every push and pull request a separate
job is run for a whole list of boards and the entire platform, including toolchain, is installed via Boards Manager,
downloading tens or hundreds of megabytes (almost entirely from downloads.arduino.cc).

You can see a demonstration of the caching system applied to the ArduinoIoTCloud library's sketch compilation workflow here:
https://github.com/per1234/ArduinoIoTCloud/commit/5e089f89786b174bbc0604a7533da309a5d60959
(it's caching directories.downloads)

Here's a run with the original workflow:
https://github.com/per1234/ArduinoIoTCloud/actions/runs/645429514

Here's a workflow run using the caches:
https://github.com/per1234/ArduinoIoTCloud/actions/runs/645607538
The caches replace about 1.5 GB of downloads.

However, you'll notice that the total duration of the non-cached workflow was 13m 31s, while the duration of the cached
workflow was 13m 56s.
I wouldn't focus on these numbers down to the second because I've noticed variation on the order of minutes between runs
of even identical workflows. However, it does seem to indicate that there isn't a dramatic difference in duration between
cached vs. non-cached.

I received some feedback that the caching efforts I made for the sketch compilation workflows were pointless since GitHub
is on AWS and arduino.cc is on AWS, which makes the download nearly instantaneous. While working on the project, I always
had it in the back of my mind that it's naive to assume that every download is transferred all the way across the
Internet from the downloads.arduino.cc server to the runner machine. But I really don't have an understanding of this
subject and I decided I should just carry on.

This makes me wonder some things:

- Why is GitHub recommending the use of caching downloaded dependencies? They provide examples of doing this with all
  major package managers, some of which are surely also on AWS.
- Why is caching done for the Python dependencies in the Arduino CLI repo?
  - https://github.com/arduino/arduino-cli/blob/414ab001a235c213f96e00f185f082783118e109/.github/workflows/publish-docs.yaml#L73-L79
  - https://github.com/arduino/arduino-cli/blob/414ab001a235c213f96e00f185f082783118e109/.github/workflows/validate-docs.yaml#L60-L66
  - https://github.com/arduino/arduino-cli/blob/414ab001a235c213f96e00f185f082783118e109/.github/workflows/link-validation.yaml#L45-L51
- Does Arduino still get charged for these inter-AWS downloads?

I think caching would be worthwhile if it results in either of:

- Reducing workflow run duration
- Reducing network usage

Since they could provide benefits:

- Improve turnover time for the developers.
- Reduce Arduino's expenses for GitHub Actions minutes (private repos) and network traffic for downloads from arduino.cc.
- General efficiency and best practices. I think we have the obligation to avoid a "tragedy of the commons" by
  unnecessarily wasting even resources that we are not directly charged for using.