Use yamllint configuration file to define the file names to check
Because there are multiple popular file types that use YAML markup, but not the standardized file extension, I previously
was using `find` to discover all the files to check with yamllint. However, it turns out that this can be defined in the
yamllint configuration file. In addition to being less prone to breakage and providing a more minimal workflow, this also
means that contributors who run yamllint locally will get the same results as the CI.