Improve the approach for communicating workflow version
Due to different project management approaches, it is sometimes necessary to provide multiple versions of a given
workflow. Previously, this was reflected in the workflow name. However, once the workflow is in use, this information is
only noise and a source for potential confusion. It's only necessary to communicate this information during the template
workflow selection process. For this reason, I am removing the information from the workflow name, leaving it only in the
documentation and metadata.