Provide instructions for command line installation of eslint and dependencies
Previously, the "Check JavaScript/TypeScript" workflows expected the user to manually merge in package.json dependencies.

The correct way to do this is via `npm install`.

Even though there is no equivalent for defining npm scripts, it seems best to just define the required scripts in the
workflow's documentation, rather than in a package.json.