Suppress excessive output from Go linting step
By default, Task echos the commands when running a task. That is usually helpful, or at least not harmful. However, in
cases where a command consists of a large amount of shell commands, it ends up being a lot of noise that makes reading
the relevant output difficult and confusing.