Improve differentiation of sketch compilation workflows for private repos
Previously, the workflow name was used to indicate whether the workflow was the variant designed for private
repositories. However, once the workflow is in place, the user has no real interest in this information, so it only
adds useless vebosity in the logs. Better is to solely differentiate it via the filename, documentation, and metadata,
leaving the workflow name stock.