Improve differentiation of sketch compilation workflows for platforms
Previously, the workflow name was used to indicate whether the workflow was the variant designed for examples of platform
bundled libraries.

However, once the workflow is in place, this information is self-evident, so it only adds useless vebosity in the logs.
Better is to solely differentiate the flavors of workflows from each via the filename, documentation, and metadata,
leaving the workflow name to explain its purpose in a concise manner within the context it will be used in.