Provide a more detailed explanation of the need for the Check License workflow
My experience indicates that this workflow is very important, and should be used in any repository that has a license
file. However, this will not be obvious to those not familiar with the license type detection system and who have not
discovered the prevalence of adulterated of license files through widescale surveys of repositories.

For this reason, it's especially important to clearly communicate the purpose for adding the file.