Remove unnecessary `schedule` trigger events
In addition to the obvious use of `schedule` trigger events in workflows, they are also important for cases where
external changes can affect the workflow run result. Without a scheduled run, this breakage will only be caught the next
time the workflow is triggered by a contribution or development work. This will be a negative experience for contributors
who assume the workflow failure was caused by something in their changes.

However, there are some disadvantages of using a schedule trigger:

- Scheduled workflows are automatically disabled for forks, which means forkers need to manually enable both workflows
  globally, as well as each scheduled workflow individually before they can run CI in their fork.
- Scheduled workflows are automatically disabled after 60 days of inactivity in a repository.
- Scheduled workflows that don't provide any useful information are wasteful of resources.

In cases where all external dependencies are pinned, and thus not likely to suffer breakage, it is not worth having a
schedule trigger solely for this purpose.