Make "Check Configuration Files" use label config schema from local path
Unlike the other JSON schemas used by the "Check Configuration Files" workflow, the one used for validating the GitHub
repository label configuration files is located in the local repository.

In addition to being silly, the previous approach of downloading the JSON schema file did not guarantee the use of the
version of the schema in sync with the state of the repository the label configuration it was validating is from.
Although the download URL could have been easily adjusted to provide that, it seems most sensible to just get the file
from the local path.