Replace speculative source URLs with current ones
The workflows and assets contain comments providing source URL to facilitate the sync of fixes and improvements between
the upstream source copies and downstream distributions. Those URLs were based on the speculation that the files might
eventually be hosted in another repository. It doesn't appear that will happen any time soon, so the comments were
misleading. It will be best to make them accurate until such time as the files might be migrated.