Sync "Sync Labels" workflow from "template"
The "Sync Labels" GitHub Actions workflow used by this repository to manage its issue/PR labels is a copy of a "template"
workflow hosted in a repository similar to this one, but scoped to the assets the Arduino Tooling Team uses in their
projects:

https://github.com/arduino/tooling-project-assets/blob/main/workflow-templates/sync-labels.md

Some enhancements and fixes have been made to the parent copy of the workflow since it was last synced. Those are pulled
in here.