Pin major version of Node.js where used directly on project in workflow templates
A best practices development strategy for a JavaScript or TypeScript-based project will define a specific version of
primary development tools such as Node.js. Previously, some of the workflow templates that used Node.js directly with
the project source did not specify a version, meaning that whatever version happens to be installed in the GitHub
Actions runner is used.

In other cases, npm is used only to manage secondary development tools (e.g., a code formatter), in which case it is
probably not so important to set a version.

The Node.js version is defined in the workflow scoped `env` key at the top of the file to make it easy to configure and
update the workflow according to the needs of the project it is applied to.