Add `eslint-plugin-jsonc` to local repo dependencies
`eslint-plugin-jsonc` is used to add JSON linting capabilities to eslint, which is used in the repository's "Check JSON"
CI workflow.

That workflow is the "standalone" style, where the focus is on CI and not on setting up for local use. In order to avoid
the addition of two extra files to the root of the repositories using the template, the plugin is installed directly in
the GitHub Actions workflow rather than the normal approach of adding it to `package.json`.

I now do have the need to use eslint locally for JavaScript, and found it was not possible due to the missing
`eslint-plugin-jsonc`:

```
$ npm run lint

> lint
> eslint --ext .js,.jsx .

Oops! Something went wrong! :(

ESLint: 8.2.0

ESLint couldn't find the plugin "eslint-plugin-jsonc".

(The package "eslint-plugin-jsonc" was not found when loaded as a Node module from the directory "E:\electronics\git-nobackup\.github".)

It's likely that the plugin isn't installed correctly. Try reinstalling by running the following:

    npm install eslint-plugin-jsonc@latest --save-dev

The plugin "eslint-plugin-jsonc" was referenced from the config file in ".eslintrc.yml".

If you still can't figure out the problem, please stop by https://eslint.org/chat/help to chat with the team.
```

So I am now installing it normally for local use.

Ideally, the workflow would also be changed over to a variant intended for use in projects managed via npm, but there is
no such template at this time so that will have to be done later.