Resolve conflict between ESLint and Prettier quoting for JS/TS
Following the predominant code style used in Arduino projects, ESLint was configured to require single quotes in
JavaScript and TypeScript.

Prettier is used as the formatter for these files and Prettier's default style is double quotes. So this resulted in
conflicting behavior between the two tools which was impossible to satisfy.

The solution is to configure Prettier to use single quotes in JavaScript and TypeScript. The Prettier configuration file
is used for this purpose, as this approach is the best suited to a "template" `.prettierrc.yml` file that can be added
to any project to provide the standard configuration.