Use 120 for Prettier `printWidth` JS/TS configuration
ESLint is used to enforce a maximum line length of 180 for JavaScript and TypeScript files. This is a hard limit, which
will result in a linting error if exceeded. For this reason, it is reasonable to make it a very generous width (based on
the original tslint configuration from Arduino Pro IDE). If the width exceeds 180, the user must investigate whether it
can be wrapped. If not, they can add a comment to disable the rule for that line.

Prettier has a different approach, where wrapping is done where possible with a target width goal, but permitting lines
to exceed that width. In this usage, a line width configuration optimized for readability is best. The Prettier default
of 80 characters is an artifact of display hardware limitations that no longer exist, so a custom configuration of 120
characters is used, consistent with the Prettier configuration used in other projects such as Arduino CLI.

So even though it might seem that the ESLint `max-len` setting of 180 and the Prettier `printWidth` configuration
settings are in conflict, in reality there is none because these are configurations for two different features which can
work symbiotically.