Resolve ESLint trailing comma configuration contradictory to Prettier
Prettier requires trailing commas on object literals in JavaScript and Typescript code.

Previously ESLint was configured to ignore the presence or absence of trailing commas. Even though this was not in
conflict with the Prettier formatting style, it is contradictory to have one style configuration explicitly indicate
that either is fine, while another mandates the comma.

The ESLint `comma-dangle` rule is disabled anyway by the `eslint-config-prettier` configuration that is in use, so this
change has no functional effect. It merely simplifies the ESLint configuration and avoids potential confusion.