Resolve ESLint block padding configuration contradiction of Prettier
Prettier prohibits leading or trailing blank lines in block in JavaScript and Typescript code.

Previously ESLint was explicitly configured to ignore the presence or absence of block padding. Even though this was not
in conflict with the Prettier formatting style, it is contradictory to have one style configuration explicitly indicate
that block padding is fine, while another prohibits it.

The ESLint `padded-blocks` rule is disabled anyway by the `eslint-config-prettier` configuration that is in use, so this
change has no functional effect. It merely simplifies the ESLint configuration and avoids potential confusion.