Remove redundant ESLint indent configuration for JS/TS
The strategy is to use Prettier as the primary code formatter for JavaScript and TypeScript, with ESLint used more for
enforcement of code quality.

Prettier already uses this indentation configuration by default, and the `eslint-config-prettier` configuration in use
will disable the ESLint `indent` rule. So having this in the ESLint configuration file only added unnecessary complexity
and maintenance burden.