Enforce comment padding in JS/TS code
Previously, ESLint was configured to downgrade the rule configuration for comment padding inherited from the Airbnb
JavaScript Style Guide from an error to a warning, following the precedent established here:

https://github.com/arduino/arduino-create-agent-js-client/blob/2.4.9/.eslintrc#L13

However, this precedent was picked at random. Unless some clear justification for making this a warning emerges, I would
prefer to keep the ESLint configuration as simple as possible.