Resolve ESLint arrow function style configuration contradiction of Prettier
Prettier requires parentheses on arrow function arguments in JavaScript and TypeScript code.

Previously ESLint was explicitly configured to ignore the presence or absence of these parentheses. Even though this was
not in conflict with the Prettier formatting style, it is contradictory to have one style configuration explicitly
indicate that either arrow function argument style is fine, while another requires a specific style.

The ESLint `arrow-parens` rule is disabled anyway by the `eslint-config-prettier` configuration that is in use, so this
change has no functional effect. It merely simplifies the ESLint configuration and avoids potential confusion.