Enforce consistent return statements in JS/TS code
Previously, ESLint was configured to disable the rule for consistency in return statements in JavaScript and TypeScript
inherited from the Airbnb JavaScript Style Guide.

This exception was added following the precedent established here:

https://github.com/arduino/arduino-create-agent-js-client/blob/2.4.9/.eslintrc#L13

However, this precedent was picked at random. Unless some clear justification for disabling this emerges, I would prefer
to keep the ESLint configuration as simple and aligned with Airbnb JavaScript Style Guide as possible.