Resolve ESLint unnecessary escaping configuration contradiction of Prettier
Prettier prohibits unnecessary escape characters in JavaScript and TypeScript code.

Previously ESLint was explicitly configured to ignore unnecessary escape characters. Even though this was not in
direct conflict with the Prettier formatting style, it is contradictory to have one style configuration explicitly
indicate that unnecessary escape characters are fine, while another prohibits them.

The ESLint `no-useless-escape` rule is disabled anyway by the `eslint-config-prettier` configuration that is in use, so
this change has no functional effect. It merely simplifies the ESLint configuration and avoids potential confusion.