Prohibit unnecessary named export in JS/TS code
Previously, ESLint was configured to disable the rule inherited from the Airbnb JavaScript Style Guide which requires
default export instead of named export when there is only a single export from a module in JavaScript and TypeScript
code.

This exception was added following the precedent established here:

https://github.com/arduino/arduino-create-agent-js-client/blob/2.4.9/.eslintrc#L15

However, this precedent was picked at random. Unless some clear justification for disabling this rule emerges, I would
prefer to keep the ESLint configuration as simple and aligned with Airbnb JavaScript Style Guide as possible.