Remove redundant ESLint type annotation spacing config for TS
The strategy is to use Prettier as the primary code formatter for TypeScript, with ESLint used more for static analysis
of the code itself.

Prettier enforces a consistent spacing around type annotations, and the `eslint-config-prettier` configuration in use
will disable the redundant ESLint rules. So having this in the ESLint configuration file only added unnecessary
complexity and maintenance burden.