Remove the "Check Dependencies" template
This template has been superseded by the one here:

https://github.com/arduino/tooling-project-assets/blob/main/workflow-templates/check-go-dependencies-task.md

Experimentation with the "License Finder" tool used by this template revealed that it has an unacceptable level of
"fuzziness" about license type detection for the intended use case. It was replaced by https://github.com/github/licensed in the finished template.