Use appropriate name for `package.json` schema temp file
A copy/paste error resulted in the temporary file path used by the `npm:validate` task for the npm `package.json` JSON
schema having "dependabot" in its name. Even though it makes no difference for this arbitrary filename from a technical
standpoint, this could result in confusion for the maintainer of the task file.