Trigger "Test C++" workflow on changes to Codecov configuration file
For the sake of efficiency, the "Test Go" GitHub Actions workflow is configured to run only when relevant files are
modified. Since the workflow uploads code coverage data to Codecov, the Codecov configuration file is one of these
relevant files.

Previously, the workflow's paths filter was not configured for the configuration file, meaning the workflow would not be
triggered on change to the Codecov configuration.

In addition to the standard Codecov configuration file name and location, an alternative filename is also recognized by
Codecov: .codecov.yml. Two subfolders are also supported in addition to the root of the repository as locations for the
configuration file.

The workflow's paths filter is hereby configured to recognize changes to any valid Codecov configuration file.