Use v1 ref of `codespell-project/actions-codespell` action in templates
The `codespell-project/actions-codespell` action is used in the GitHub Actions workflows as a convenient way to run the
codespell spell checker tool on the files of the repository.

At the time the template workflows were created, no release had yet been made of the action. For this reason, we used
the `master` ref in the workflows, which caused them to use the action at the development state from the tip of the
`master` branch of the action's repository. This approach is not conducive to stability of the project infrastructure,
but no other option was available. A release of the action has now been made. To accompany it, a major version ref has
been added to that repository. It will always point to the latest release of the "1" major version series.

Use of the major version ref will cause the workflow to use a stable version of the action, while also benefiting from
ongoing development to the action up until such time as a new major release of an action is made. At that time we would
need to evaluate whether any changes to the workflow are required by the breaking change that triggered the major
release before manually updating the major ref (e.g., uses: `codespell-project/actions-codespell@v2`). I think this
approach strikes the right balance between stability and maintainability for these workflows.