Check for outdated actions in workflow templates in place
Dependabot supports managing GitHub Actions action dependencies of workflows. In addition to the workflows used in the
repository's own infrastructure, this system was also used to maintain the dependencies of the workflow templates.

At the time it was set up, Dependabot had an inconvenient behavior where workflows were expected to be found in the
`.github/workflows/` subfolder of the path specified via the `directory` key of the `dependabot.yml` configuration
file. This meant it was not possible to directly scan the workflows in the `workflow-templates` folder. The workaround
was to put a set of copies in `workflow-templates/dependabot/workflow-template-copies/.github/workflows/`. The
Dependabot PRs updating those workflows were not merged, but instead used as a notification of available updates for the
source workflows in `workflow-templates`.

A change in Dependabot in spring of 2022 caused the system to stop working. That bug has been fixed. During the fix, the
odd system of not looking for workflows in the folder specified via the `directory` key was dropped, which means it is
now possible to use Dependabot to directly maintain the workflow templates, making the duplicates no longer necessary.