Remove outdated comment from Dependabot config file
Dependabot supports managing GitHub Actions action dependencies of workflows. In addition to the workflows used in the
repository's own infrastructure, this system is also used to maintain the dependencies of the workflow templates. This
is done by adding an additional `github-actions` element to the Dependabot configuration file.

At the time that system was originally set up, Dependabot had an unintuitive behavior where workflows were expected to
be found in the `.github/workflows/` subfolder of the path specified via the `directory` key of the `dependabot.yml`
configuration file. A comment was added to the configuration file to explain this.

Since then, the odd behavior was dropped and the actual path specified via the `directory` key is used. The comment is
now incorrect and therefore must be removed.