Use non-executable shell script check code from arduino/tooling-project-assets
There is a Task-based variant of the "Check Shell" template workflow in the arduino/tooling-project-assets repository.

That code is designed for the additional use case of running locally on the contributor machine, which is not the case
for this template where the code only ever runs in a workflow. This results in some of the code from the task not being
applicable for this use case, though it is still perfectly functional. Despite this incongruity, I think it will be
beneficial to maintain only a single version of the code for use in both templates.