Prevent "Report Size Deltas" workflow runs in private repos
The `arduino/report-size-deltas` GitHub Actions action is used to report changes in sketch memory usage that would
result from merging a PR.

In order to avoid excess GitHub service charges, a different approach must be taken when using this action in a private
repository compared to a public repository.

Because it runs on a frequent schedule, the "Report Size Deltas" workflow uses a large quantity of GitHub Actions
minutes. Public repositories have unlimited free minutes, so this is not a concern in those projects, but minutes are
charged in private repositories, so this workflow should not be used in private repositories.

Even though this is clearly documented in the installation instructions and a separate set of workflow templates
provided for use in private repositories, it still happens that the workflow is added to private repos by accident. So
it will be safest to configure the workflow to only run when it is in a public repository.