Remove unnecessary `contents: read` permission from public variants of sketch compilation workflows
The `permission` key is used to configure which permissions are granted to the automatically generated GITHUB_TOKEN
access token when the workflow job runs in a trusted context.

In private repositories, the `contents: read` permission is required for the workflow job to have access to the
repository contents in private repositories, including checkout. Although this is not required for public repositories,
the templates are generally intended to be usable in all repositories so the `contents: read` permission is granted even
though this means a slight excess of permissions when the workflow is installed in a public repository.

The situation is different when it comes to the sketch compilation workflows because in this case it is already
necessary to maintain separate variants of the workflows for use in public and private repositories. For this reason,
there is no point in granting the `contents: read` permission in the variants for use in public repositories.