Use correct syntax for empty YAML sequence in `libraries` input of `arduino/compile-sketches` step
Unlike the `arduino/compile-sketches` action's other inputs (`platforms`, `sketch-paths`) that take a "sequence" (AKA
array, list) type YAML document as input, because a sketch may have no library dependencies it is valid and common to
pass an empty sequence via the `libraries` input.

Due to the need to maintain backwards compatibility with the original short-sighted design of the action API, which
didn't consider the action being used with any project type other than libraries, it is necessary for the user to do
this in their workflow rather than simply omitting the `libraries` input.

Previously "-" was used for this purpose. That is supported by the action, but it is not semantically accurate. "-" is
not an empty sequence, but rather a sequence consisting of a single null element. The only way to specify an empty
sequence in YAML is `[]`.