Avoid installation of platform as library in platform sketch compilation workflows
Due to an original short-sighted design, which didn't consider any project type other than libraries, the default
behavior of the `arduino/compile-sketches` action is to install the repository as a library.

Previously the platform example compilation workflows used the default behavior, which caused a copy of the platform to
be installed as a library. Since that would never be done in a real world environment, this could possibly cause the
results of the workflow to be inaccurate (e.g., spurious compilation errors or warnings), and also introduces
inefficiencies.

This default behavior is overridden when library dependencies are configured via the action's `libraries` input. In
cases where there are no library dependencies, an YAML document containing only an empty sequence can be used in the
input.