Adjust spell checker configuration to fix false positive
The codespell tool is used for automated detection of common misspellings in the files of this project.

A new release of codespell includes an expansion of its misspellings dictionary. It now detects several misspelled words in the SVG files in the repository. Those files are not maintained in this repository so it is better to configure codespell to ignore them altogether, which is done by adding the .svg file extension to the list of files to skip in the codespell configuration.