Remove unnecessary GITHUB_TOKEN permission from "Sync Labels" workflow
In keeping with the security "principle of least privilege", the GitHub Actions workflows are configured to only give
the GITHUB_TOKEN token the permissions the workflow requires.

The "Sync Labels" workflow must be able to use the token to adjust the labels in issues and pull requests in the
repository. For this reason, the token was configured to have `issues: write` and `pull-requests: write` permissions.

It turns out that the `issues: write` permission alone provides this capability for pull requests as well as issues and
the `pull-requests: write` permission is completely unnecessary. The reason is that the GitHub API considers all pull
requests to also be an issue.