Don't upload multiple times to same artifact in sketch compilation workflows
The sketch compilation GitHub Actions workflows are configured to compile the sketches for each of a set of supported
boards. This is done by using a job matrix in the GitHub Actions workflow to generate a parallel job for each board.

A GitHub Actions workflow artifact is used to transfer the sketches report files generated by the
"arduino/compile-sketches" action between this workflow's job and the "Report Size Deltas" workflow that uses the
"arduino/report-size-deltas" action to publish the data. The "actions/upload-artifact" action is used to upload the
sketches report files to the workflow artifact.

Previously, the sketches reports from all the boards were uploaded to a single artifact. However, support for uploading
multiple times to a single artifact was dropped in version 4.0.0 of the "actions/upload-artifact" action. So it is now
necessary for each of the jobs to use a separate artifact. The "arduino/report-size-deltas" action is configured to
recognize these multiple artifacts by a regular expression that matches against a standardized prefix on the artifact
names.